requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
httpx
tenacity
duckduckgo-search
beautifulsoup4
readability-lxml
prompt_toolkit
rich
chromadb
sentence-transformers
pypdf
numpy
scikit-learn
sqlalchemy
